requirements.txt
GitPython
radon
pylint
requests
streamlit
